requirements.txt
Django==2.0.4
mysqlclient==1.3.12
requests==2.18.4
beautifulsoup4==4.6.0
lxml==4.2.1
djangorestframework==3.8.2